Move all Python code into a src directory
Theres many, many reasons why this is done. It helps cut down the root directory in-place for config files, dotfiles, automation files, and more. It also helps for splitting stuff like source from documentation, as well as automation.

But most importantly, it makes it a python-package, for easier importing across the codebase between each other. It also makes it easier to reference code.